Significant rewrite of ClientInterfacer.
The new code is much simpler, by making use of a new invariant: there
are never commands in the queue if there's room to send them to the
server. Also, I think the functions are ordered a little better.

Tweak a couple scripts to work with the new, slightly different API.