Quick-and-dirty port of the important part of travel.py.
This is just enough of a conversion so that I can run the new script to
walk between predefined waypoints; there's no way to record new paths.
It makes little difference since in practice I only really use this to
walk between Scorn and Navar (and maybe Scorn and Lake Country), which
path is already defined.