Add partial support for "special" commands.
These are the sorts of commands that a normal player would issue by
clicking on particular items, rather than typing some text into the
command field. They take item tags, so aren't subject to the problem
where you use some string to identify an item but it actually matches
more than one item.

Unfortunately, apparently these commands don't give you a "watch comc",
which means that timing them properly is going to be tricky.

Also add an in-progress script to make water of the wise using a
cauldron on the ground at your feet. It's not finished because the
special commands aren't really in a usable state yet.