Finish implementing workaround for non-getsComc commands.
The main change here is that if there are too many non-getsComc commands
in a row in the queue, then we intersperse "stay", which is a no-op that
gets a "comc" acknowledgement, so that we can wait for only a few of the
non-getsComc commands at a time.

Add a basic test of this new functionality in test_move.py, and update
water_of_the_wise.py to actually work now that I've (AFAIK) properly
implemented the functionality it depends on.