Further progress on map parsing.
parseRoom could be done, given the right definition of the Room class
and modulo any bugs I find when I actually run it. (And also discounting
the fact that it doesn't yet support custom path markers.)

Probably once I actually try to implement the Room class for these
implicit requirements, I'll end up changing the interface around a bit,
but so it goes.